implement bit watch extension by re-using most of the old bit-watch code and using "compile" extension as a dependency
