+++ src/constants.ts
@@ -419,3 +419,7 @@ export const CONCURRENT_IO_LIMIT = 100; // limit number of files to read/write/d
 export const UNABLE_TO_LOAD_EXTENSION = (id: string) => `couldn't load extension ${id}, see full error in the log file`;
 export const UNABLE_TO_LOAD_EXTENSION_FROM_LIST = (ids: string[]) =>
   `couldn't load one of the following extensions ${ids.join(', ')}, see full error in the log file`;
+
+// todo: move the following two lines to the watch extension once its e2e moved to the extension dir
+export const STARTED_WATCHING_MSG = 'started watching for component changes to rebuild';
+export const WATCHER_COMPLETED_MSG = 'watching for changes';
