+++ src/extensions/watch/watch.cmd.tsx
@@ -1,17 +1,22 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 import React from 'react';
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { Color } from 'ink';
-import { Command } from '../paper';
+import { Command, CLIArgs, PaperOptions } from '../paper';
+import { Watch } from '.';
 
-export default class ServeCommand implements Command {
-  name = 'watch [id...]';
+export class WatchCommand implements Command {
+  name = 'watch';
   description = 'watch a set of components';
   alias = '';
   group = '';
   shortDescription = '';
-  options = [];
+  options = [['v', 'verbose', 'showing npm verbose output for inspection and prints stack trace']] as PaperOptions;
 
-  async render() {
-    return <Color>hi amit</Color>;
+  constructor(private watch: Watch) {}
+
+  async render(cliArgs: CLIArgs, { verbose = false }: { verbose?: boolean }) {
+    await this.watch.watch({ verbose });
+    return <Color>watcher terminated</Color>;
   }
 }
