+++ src/extensions/watch/watch.provider.ts
@@ -1,6 +1,14 @@
+import { Workspace } from '../workspace';
+import { BitCli } from '../cli';
+import { WatchCommand } from './watch.cmd';
+import { Compile } from '../compile/compile';
 import Watch from './watch';
-import LegacyWatch from '../../consumer/component-ops/watch-components';
 
-export function provideWatch() {
-  return new Watch(new LegacyWatch(true));
+export type watchDeps = [BitCli, Compile, Workspace];
+
+export function provideWatch([cli, compile, workspace]: watchDeps) {
+  const watch = new Watch(compile, workspace);
+  // @ts-ignore
+  cli.register(new WatchCommand(watch));
+  return watch;
 }
